fix(layout): serve Organization JSON-LD from static shell with CSP hash
Remove the per-request headers() boundary that kept public routes fully
dynamic so ISR segments can be prerendered after CP08-03/06.
