+++ apps/site/src/components/layouts/site-static-shell.tsx
@@ -1,6 +1,6 @@
 import { fontClassNames } from '@/lib/font';
 
-import { SiteHeadNonceScripts } from '@/components/layouts/site-head-nonce-scripts';
+import { SiteOrganizationSchema } from '@/components/layouts/site-organization-schema';
 
 const criticalCSS = `
   /* Essential design tokens for critical styles */
@@ -50,7 +50,7 @@ export function SiteStaticShell({ children }: Readonly<{ children: React.ReactNo
     <html lang="en" className={fontClassNames} suppressHydrationWarning>
       <head>
         <style dangerouslySetInnerHTML={{ __html: criticalCSS }} suppressHydrationWarning />
-        <SiteHeadNonceScripts />
+        <SiteOrganizationSchema />
       </head>
       <body className="flex flex-col min-h-screen">{children}</body>
     </html>
